requirements.txt
numpy==1.23.3
vpython==7.6.4
pyserial==3.5